remove asteroidInterfaces
