+++ src/config/driverInterfaces.ts
@@ -43,3 +43,12 @@ export interface ILogger {
 export interface ICallback {
   (error: Error | null, ...args: any[]): void
 }
+
+/** Credentials for login method */
+export interface ICredentials {
+  password: string,
+  username?: string,
+  email?: string,
+  ldap?: boolean,
+  ldapOptions?: object
+}
